crane: rename asr.rs to asr_client.rs
Frees up asr.rs for the upcoming Asr trait, mirroring how tts.rs holds
the Tts trait while tts_client.rs holds the legacy TtsClient.
